requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
Django==4.2
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
pytz==2023.3
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
